slack/requirements.txt
requests==2.22.0
slackclient==2.5.0
python-dotenv==0.10.5

